# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NPCs AI Studio · 本地数据服务",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/templates": {
      "get": {
        "summary": "Template Library",
        "operationId": "template_library_api_templates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Capture Template",
        "operationId": "capture_template_api_templates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureTemplate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/templates/{template_id}/archive": {
      "post": {
        "summary": "Archive Template",
        "operationId": "archive_template_api_templates__template_id__archive_post",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveTemplate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/projects/{project_id}/handoff": {
      "post": {
        "summary": "Handoff",
        "operationId": "handoff_api_v2_projects__project_id__handoff_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/files/open": {
      "post": {
        "summary": "Open File",
        "operationId": "open_file_api_files_open_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/new": {
      "post": {
        "summary": "New File",
        "operationId": "new_file_api_files_new_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiskProjectInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/projects/{project_id}/save": {
      "post": {
        "summary": "Save File",
        "operationId": "save_file_api_v2_projects__project_id__save_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/projects/{project_id}/recovery": {
      "get": {
        "summary": "Recovery",
        "operationId": "recovery_api_v2_projects__project_id__recovery_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/projects/{project_id}/restore": {
      "post": {
        "summary": "Restore",
        "operationId": "restore_api_v2_projects__project_id__restore_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/projects/validate": {
      "post": {
        "summary": "Validate",
        "operationId": "validate_api_v2_projects_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/projects/{project_id}": {
      "get": {
        "summary": "Get",
        "operationId": "get_api_v2_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/projects/{project_id}/commands": {
      "post": {
        "summary": "Commands",
        "operationId": "commands_api_v2_projects__project_id__commands_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandBatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/projects/{project_id}/edit-history": {
      "get": {
        "summary": "Edit History",
        "operationId": "edit_history_api_v2_projects__project_id__edit_history_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Undo Edit",
        "operationId": "undo_edit_api_v2_projects__project_id__edit_history_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArchiveTemplate": {
        "properties": {
          "library_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Library Revision"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "library_revision",
          "archived"
        ],
        "title": "ArchiveTemplate"
      },
...
      "CaptureTemplate": {
        "properties": {
          "project_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$",
            "title": "Project Id"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "library_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Library Revision"
          },
          "source_kind": {
            "type": "string",
            "enum": [
              "character",
              "dialogue"
            ],
            "title": "Source Kind"
          },
          "source_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$",
            "title": "Source Id"
          },
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "expected_revision",
          "library_revision",
          "source_kind",
          "source_id",
          "name"
        ],
        "title": "CaptureTemplate"
      },
...
      "DiskProjectInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name"
          },
          "folder": {
            "type": "string",
            "title": "Folder"
          },
          "filename": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Filename"
          },
          "template": {
            "type": "string",
            "enum": [
              "blank",
              "lighthouse",
              "outpost",
              "campfire"
            ],
            "title": "Template",
            "default": "blank"
          },
          "layout": {
            "type": "string",
            "enum": [
              "file",
              "folder"
            ],
            "title": "Layout",
            "default": "file"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "folder",
          "filename"
        ],
        "title": "DiskProjectInput"
      },
      "DocumentInput": {
        "properties": {
          "document": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Document"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document"
        ],
        "title": "DocumentInput"
      },
...
      "ExportRequest": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "format": {
            "type": "string",
            "enum": [
              "project",
              "markdown",
              "csv"
            ],
            "title": "Format"
          },
          "level_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "format"
        ],
        "title": "ExportRequest"
      },
...
      "FileInput": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "discard": {
            "type": "boolean",
            "title": "Discard",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "path"
        ],
        "title": "FileInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryInput": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "direction": {
            "type": "string",
            "enum": [
              "undo",
              "redo"
            ],
            "title": "Direction"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "direction"
        ],
        "title": "HistoryInput"
      },
...
      "JsonValue": {},
...
      "Metadata": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "migration": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MigrationInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Metadata"
      },
      "MigrationInfo": {
        "properties": {
          "source_version": {
            "type": "integer",
            "const": 1,
            "title": "Source Version"
          },
          "warnings": {
            "items": {
              "type": "string",
              "maxLength": 30000
            },
            "type": "array",
            "title": "Warnings"
          },
          "legacy_preview_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LegacyPreviewSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "legacy_world_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legacy World Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_version"
        ],
        "title": "MigrationInfo"
      },
...
      "Project": {
        "properties": {
          "format": {
            "type": "string",
            "const": "ludo-npc-project",
            "title": "Format",
            "default": "ludo-npc-project"
          },
          "schema_version": {
            "type": "integer",
            "const": 2,
            "title": "Schema Version",
            "default": 2
          },
          "project_id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,99}$",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Name",
            "default": "未命名项目"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision",
            "default": 1
          },
          "content_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Content Revision",
            "default": 1
          },
          "layout_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Layout Revision",
            "default": 1
          },
          "content": {
            "$ref": "#/components/schemas/Content"
          },
          "editor": {
            "$ref": "#/components/schemas/Editor"
          },
          "metadata": {
            "$ref": "#/components/schemas/Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Project"
      },
...
      "RestoreInput": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "point_id": {
            "type": "string",
            "title": "Point Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "point_id"
        ],
        "title": "RestoreInput"
      },
...
      "SaveInput": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision"
        ],
        "title": "SaveInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}